+++ packages/browser-sdk/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @phantom/browser-sdk
 
+## 1.0.0-beta.20
+
+### Patch Changes
+
+- c3f5969: Do not reconnect automatically on injected provider if user disconnected
+- 462c2b5: Get authUserId from injected provider
+  - @phantom/client@1.0.0-beta.20
+  - @phantom/embedded-provider-core@1.0.0-beta.20
+
+## 1.0.0-beta.19
+
+### Patch Changes
+
+- d35daaa: Bump all dependencies
+- Updated dependencies [d35daaa]
+  - @phantom/embedded-provider-core@1.0.0-beta.19
+  - @phantom/client@1.0.0-beta.19
+  - @phantom/base64url@1.0.0-beta.8
+  - @phantom/constants@1.0.0-beta.8
+  - @phantom/sdk-types@1.0.0-beta.8
+  - @phantom/chain-interfaces@1.0.0-beta.8
+  - @phantom/parsers@1.0.0-beta.8
+  - @phantom/browser-injected-sdk@1.0.0-beta.6
+  - @phantom/indexed-db-stamper@1.0.0-beta.2
+
+## 1.0.0-beta.18
+
+### Patch Changes
+
+- f1f401d: Support authuser id from wallet
+- 0ee572a: EVM SDKs
+- Updated dependencies [0ee572a]
+  - @phantom/embedded-provider-core@1.0.0-beta.18
+  - @phantom/client@1.0.0-beta.18
+
+## 1.0.0-beta.17
+
+### Patch Changes
+
+- bd2219e: Fixed features access
+  - @phantom/client@1.0.0-beta.17
+  - @phantom/embedded-provider-core@1.0.0-beta.17
+
 ## 1.0.0-beta.16
 
 ### Patch Changes